refactor: review-sandbox and dev-sandbox to share common commands and logic
- Refactor `cmd/review-sandbox/main.go` to use `cobra` for argument parsing and subcommand support, aligning it with `dev-sandbox`.
- Extract review agent logic into `pkg/commands/review.go`.
- Extract SSHD command logic into `pkg/commands/sshd.go` as `BuildSSHDCommand` and integrate it into both sandboxes.
- Extract CodeServer command logic into `pkg/commands/codeserver.go` as `BuildCodeServerCommand` and integrate it into both sandboxes.
- Move `ImageBuilder` logic from `cmd/dev-sandbox` to `pkg/imagebuilder` to promote reuse and cleanup `main.go`.
- Update `pkg/codeserver` to support repository-level URLs in addition to PR URLs.
- Rename command constructors to `Build*Command` for consistency with existing patterns.
