+++ repo-agent/pkg/imagebuilder/imagebuilder.go
@@ -1,0 +1,90 @@
+package imagebuilder
+
+import (
+	"context"
+	"fmt"
+	"os"
+	"os/exec"
+	"path/filepath"
+	"strings"
+
+	"k8s.io/klog/v2"
+)
+
+// ImageBuilder is responsible for initializing the container, e.g. installing dotfiles
+type ImageBuilder struct {
+}
+
+// InstallDotfilesRepo clones and install a dotfiles repo
+func (b *ImageBuilder) InstallDotfilesRepo(ctx context.Context, dotFilesRepo string) error {
+	log := klog.FromContext(ctx)
+
+	// Get the user's cache directory
+	cacheDir, err := os.UserCacheDir()
+	if err != nil {
+		return fmt.Errorf("getting user cache dir: %w", err)
+	}
+
+	dotfilesDir := filepath.Join(cacheDir, "dev-sandbox", "dotfiles")
+
+	if err := b.GitClone(ctx, dotFilesRepo, dotfilesDir); err != nil {
+		return err
+	}
+
+	// Well-known entrypoints
+	entrypoints := []string{
+		"setup",
+	}
+
+	var foundEntrypoint string
+	for _, entrypoint := range entrypoints {
+		p := filepath.Join(dotfilesDir, entrypoint)
+		if _, err := os.Stat(p); err != nil {
+			if !os.IsNotExist(err) {
+				log.Error(err, "error checking for entrypoint", "entrypoint", p)
+			}
+			continue
+		}
+		foundEntrypoint = p
+		break
+	}
+
+	if foundEntrypoint == "" {
+		return fmt.Errorf("unable to find entrypoint in dotfiles repo %q", dotFilesRepo)
+	}
+
+	cmd := exec.CommandContext(ctx, foundEntrypoint)
+	cmd.Stdout = os.Stdout
+	cmd.Stderr = os.Stderr
+
+	if err := cmd.Run(); err != nil {
+		return fmt.Errorf("error running dotfiles entrypoint %q from repo %q: %w", strings.Join(cmd.Args, " "), dotFilesRepo, err)
+	}
+
+	return nil
+}
+
+// GitClone clones a git repo to the dest directory.
+func (b *ImageBuilder) GitClone(ctx context.Context, source string, dest string) error {
+	log := klog.FromContext(ctx)
+
+	args := []string{
+		"git",
+		"clone",
+		source,
+		dest,
+	}
+
+	cmdString := strings.Join(args, " ")
+	log.Info("cloning git repo", "source", source, "command", cmdString)
+
+	cmd := exec.CommandContext(ctx, args[0], args[1:]...)
+	cmd.Stdout = os.Stdout
+	cmd.Stderr = os.Stderr
+
+	if err := cmd.Run(); err != nil {
+		return fmt.Errorf("cloning git repo %q with %q: %w", source, cmdString, err)
+	}
+
+	return nil
+}
